fixes #344
